feat: add not found middlewares

+++ src/services/orders.service.ts
@@ -447,6 +447,16 @@ export const listAllRefusedService = async () => {
 };
 
 export const deleteAnOrderService = async (id: string) => {
+  const toDelete = await prisma.orders.findFirst({
+    where: {
+      id,
+    },
+  });
+
+  if (!toDelete) {
+    throw new AppError('Order not found', 404);
+  }
+
   await prisma.orders.delete({
     where: {
       id,
